Implement grapple physics and linking code
