+++ src/main/java/jard/alchym/client/ExtraPlayerDataAccess.java
@@ -3,8 +3,13 @@ package jard.alchym.client;
 import net.minecraft.client.model.ModelPart;
 import net.minecraft.util.math.Vec3d;
 
+import java.util.Stack;
+
 public interface ExtraPlayerDataAccess {
     Vec3d getPrevVel ();
-    ModelPart getCloak ();
+    Stack<Vec3d> getGrapple ();
+    double getGrappleLength ();
     boolean isJumping ();
+
+    ModelPart getCloak ();
 }
